requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
backrefs==5.9
certifi==2025.6.15
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distlib==0.3.9
distro==1.9.0
docstring_parser==0.16
exceptiongroup==1.3.0
filelock==3.18.0
frozenlist==1.7.0
ghp-import==2.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.12
idna==3.10
iniconfig==2.1.0
instructor==1.9.0
Intercom==0.2.3
Jinja2==3.1.6
jiter==0.10.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.15
mkdocs-material-extensions==1.3.1
multidict==6.6.3
nodeenv==1.9.1
openai==1.93.0
packaging==25.0
paginate==0.5.7
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.2.0
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pymdown-extensions==10.16
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
pyyaml_env_tag==1.1
pyzmq==27.0.0
requests==2.32.4
rich==14.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
tenacity==9.1.2
tomli==2.2.1
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
virtualenv==20.31.2
watchdog==6.0.0
yarl==1.20.1
